Create complete_set
This is the base dataset used for the article "Somehow Nothing Drew Me into Novyi Mir." It consists of entries for all contributions to the journals discussed over a given time period - compiled from the digitized Letopis' zhurnal'nykh statei, parsed, and checked for completion. For the data in this file, the only (few) missing contributions are those that were accidentally skipped in the Letopis' or that cannot be located in the Letopis' due to typos or other misrepresentations.