Add table of journal issues and their circulation
This is a tsv file, listing the print run (i.e. number of copies printed) for each issue of the six journals (Znamia, Molodaia gvardiia, Nash sovremennik, Novyi mir, Oktiabr', Iunost') for the time period 1958-1972.
It features the columns 'journal' (journal title), 'issue' (issue number, out of 1-12 or 1-6), 'year' (year of the issue), 'print_run' (circulation of the journal according to information given in the back matter of the journal issue), 'issues_pa' (number of issues published during the given year - 12 = monthly, 6 = bimonthly), 'notes' (notes on how information was captured.